+++ src/irdevice/airconditioner.ts
@@ -314,7 +314,7 @@ export class AirConditioner extends irdeviceBase {
   async pushChanges(bodyChange: any): Promise<void> {
     await this.debugLog('pushChanges')
     if (this.device.connectionType === 'OpenAPI' && !this.disablePushDetail) {
-      await this.infoLog(`Sending request to SwitchBot API, body: ${bodyChange},`)
+      await this.infoLog(`Sending request to SwitchBot API, body: ${JSON.stringify(bodyChange)}`)
       try {
         const { body } = await this.pushChangeRequest(bodyChange)
         const deviceStatus: any = await body